+++ backend/database/db.js
@@ -2,11 +2,17 @@ require("dotenv").config()
 const { Pool } = require('pg')
 
 
-console.log(process.env.DB_URL);
+if (process.env.NODE_ENV == "test") {
+    console.log("test DB")
+    db = new Pool({
+        connectionString: process.env.TEST_DB_URL
+    })
+} else {
+    db = new Pool({
+        connectionString: process.env.DB_URL
+    })
+}
 
-const db = new Pool({
-	connectionString: process.env.DB_URL
-})
-
+console.log("DB connection established")
 
 module.exports = db
